Multiplayer: Separated client from server
